+++ gui/screens/planet_number.py
@@ -21,10 +21,13 @@ class PlanetNumberScreen(TkinterApp):
     def initial_widgets(self):
         self.planet_number_entry = Entry(self.context.app)
         self.time_scale_entry = Entry(self.context.app)
-        self.pop_button = Button(self.context.app, text='back', height=1, width=9, command=self.back)
-        self.label1 = Label(self.context.app, text="planet number : ", width=12)
+        self.pop_button = Button(
+            self.context.app, text='back', height=1, width=9, command=self.back)
+        self.label1 = Label(
+            self.context.app, text="planet number : ", width=12)
         self.label2 = Label(self.context.app, text="time scale : ", width=12)
-        self.button = Button(self.context.app, text='next', height=1, width=15, command=self.next)
+        self.button = Button(self.context.app, text='next',
+                             height=1, width=15, command=self.next)
 
     def initial_position(self):
         self.pop_button.place(relx=.2, rely=.1, anchor=CENTER)
@@ -37,12 +40,14 @@ class PlanetNumberScreen(TkinterApp):
     def show_error(self):
         self.planet_number_entry.delete(0, 'end')
         self.time_scale_entry.delete(0, 'end')
-        messagebox.showerror("Error", "enter a valid number between 1 and 100 please")
+        messagebox.showerror(
+            "Error", "enter a valid number between 1 and 100 please")
 
     def next(self):
         try:
             planet_number = int(self.planet_number_entry.get())
-            self.context.environment.time_scale = int(self.time_scale_entry.get())
+            self.context.environment.time_scale = float(
+                self.time_scale_entry.get())
             self.context.environment.set_time_speed(1)
         except ValueError:
             self.show_error()
